% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autofilt.R
\name{autofilt}
\alias{autofilt}
\title{Automatic EMG filtering}
\usage{
autofilt(
  input_data,
  filter_type,
  fs,
  order,
  ftype,
  high_cut,
  low_cut,
  rp,
  rs,
  preview_plot = TRUE
)
}
\arguments{
\item{input_data}{A numeric vector representing the raw EMG data.}

\item{filter_type}{A character string specifying the type of filter design. Options include: 'butter' (Butterworth), 'cheb1' (Chebyshev type 1), 'cheb2' (Chebyshev type 2), and 'ellip' (Elliptic Filter).}

\item{fs}{Sampling rate of the EMG data.}

\item{order}{Filter order, needs to be an integer scalar input.}

\item{ftype}{A character string specifying the filter type. Options include: 'high' (highpass), 'low' (lowpass), and 'pass' (bandpass).}

\item{high_cut}{High cut frequency for highpass and bandpass filters. Required argument if ftype is 'high' or 'band'.}

\item{low_cut}{Low cut frequency for lowpass and bandpass filters. Required argument if ftype is 'low' or 'band'.}

\item{rp}{Passband ripple for Chebyshev and Elliptic filters.}

\item{rs}{Stopband attenuation for Elliptic filters.}

\item{preview_plot}{Logical, if TRUE (default), running the function returns a plot of the filtered data. If set to FALSE, no figure is generated.}
}
\value{
This function returns a numeric vector that contains the filtered EMG data as well as a plot of the filtered EMG data.
}
\description{
\code{autofilt} is an all-in-one EMG filtering function that designs and implements IIR filters in a single function call. The function designs and implements commonly used IIR filters including Butterworth, Chebyshev (type 1 and 2), and Elliptic filters. Additionally, the function returns a summary figure with the filtered EMG plotted on top of the raw EMG.
}
\examples{
\dontrun{
# Example usage:
data(EMG_data)

filt_butter <- autofilt(
  input_data = EMG_data,
  filter_type = 'butter',
  fs = 5000,
  order = 2,
  ftype = 'pass',
  high_cut = 1000,
  low_cut = 100
)
  
filt_cheby1 <- autofilt(
  input_data = EMG_data,
  filter_type = 'cheb1',
  fs = 5000,
  order = 2,
  ftype = 'pass',
  high_cut = 1000,
  low_cut = 100,
  rp = 0.5
)
  
filt_cheby2 <- autofilt(
  input_data = EMG_data,
  filter_type = 'cheb2',
  fs = 5000,
  order = 2,
  ftype = 'pass',
  high_cut = 1000,
  low_cut = 100,
  rp = 0.5
)
  
filt_ellip <- autofilt(
  input_data = EMG_data,
  filter_type = 'ellip',
  fs = 5000,
  order = 2,
  ftype = 'pass',
  high_cut = 1000,
  low_cut = 100,
  rp = 0.5,
  rs = 0.5
)

}
}
\author{
Ethan Benevides
}
