% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandpower.R
\name{bandpower}
\alias{bandpower}
\title{Calculate the Power in Specific Frequency Bands}
\usage{
bandpower(input_data, fs, freq_band)
}
\arguments{
\item{input_data}{A numeric vector representing the EMG data.}

\item{fs}{Sampling rate of the EMG data.}

\item{freq_band}{Numeric vector specifying the frequency band of interest. It can either be a single frequency or a range defined by two values.}
}
\value{
A numeric scalar value representing the average power of the specified frequency/frequency band.
}
\description{
The \code{bandpower} function calculates the power of frequency components within a specified frequency band using the Fast Fourier Transform (FFT) algorithm.
}
\examples{
\dontrun{
# Example usage:
data(EMG_data)

bandpower_output <- bandpower(input_data = EMG_data, fs = 5000, freq_band = c(250,450)))
}
}
\author{
Ethan Benevides
}
