% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMG_data.R
\docType{data}
\name{EMG_data}
\alias{EMG_data}
\title{Example EMG data recorded from the diaphragm of a mouse}
\format{
A single numeric vector representing EMG data.
}
\usage{
EMG_data
}
\description{
A small, example EMG dataset recorded from the diaphragm muscle of an anesthetized mouse.
Included here to help illustrate the use of the EMGhelpR package.
}
\details{
These data were sampled at 5000 samples per second.
}
\keyword{datasets}
