% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakfinder.R
\name{peakfinder}
\alias{peakfinder}
\title{Find Local Maxima in EMG Data}
\usage{
peakfinder(input_data, threshold = 0, min_distance = 0, preview_plot = TRUE)
}
\arguments{
\item{input_data}{A numeric vector representing EMG data.}

\item{threshold}{A numeric value representing the minimum amplitude required for a point to be considered a peak (default is 0).}

\item{min_distance}{A numeric value representing the minimum distance (in samples) between consecutive peaks (default is 0).}

\item{preview_plot}{Logical, indicating whether to display a preview plot of EMG data with detected peaks (default is TRUE).}
}
\value{
A list containing two numeric vectors: \code{peaks} representing the amplitudes of the detected peaks, and \code{locs} representing the corresponding sample indices.
}
\description{
The \code{peakfinder} function identifies local maxima (peaks) in electromyography (EMG) data. This function returns the values of the detected peaks and the corresponding sample indices. Additionally, a plot displaying the EMG data and detected peaks is automatically generated.
}
\examples{
\dontrun{
# Example usage:
data(EMG_data)

peak_out <- peakfinder(input_data = EMG_data, threshold = 0.1, min_distance = 2500)
}
}
\author{
Ethan Benevides
}
