% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa.R
\name{psa}
\alias{psa}
\title{Power Spectral Density Analysis}
\usage{
psa(input_data, fs, preview_plot = TRUE)
}
\arguments{
\item{input_data}{A numeric vector representing the EMG data.}

\item{fs}{Sampling rate of the EMG data.}

\item{preview_plot}{Logical, indicating whether to display a plot of frequency vs power (default is TRUE).}
}
\value{
A list containing two numeric vectors: frequency vector (\code{freq_vec}) and corresponding power spectral density vector (\code{power_vec}).
}
\description{
The \code{psa} function calculates the power spectral density of electromyography (EMG) data using the Fast Fourier Transform (FFT) algorithm. Additionally, it returns a plot of frequency vs power for the analyzed EMG trace.
}
\examples{
\dontrun{
# Example usage:
data(EMG_data)

psa_output <- psa(input_data = EMG_data, fs = 5000)
}
}
\author{
Ethan Benevides
}
