% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recint.R
\name{recint}
\alias{recint}
\title{Rectify and Integrate EMG Data}
\usage{
recint(input_data, window, fs, time_units = "samples", preview_plot = TRUE)
}
\arguments{
\item{input_data}{A numeric vector representing the EMG data.}

\item{window}{The size of the integration window.}

\item{fs}{Sampling rate of the EMG data.}

\item{time_units}{A character string specifying the time units for the window size. Options include: 'samples' (default), 'ms' (milliseconds), and 'sec' (seconds).}

\item{preview_plot}{Logical, indicating whether to display a preview plot of the rectified and integrated EMG data (default is TRUE).}
}
\value{
A numeric vector representing the rectified and integrated EMG data as well as a plot of the rectified and integrated EMG data.
}
\description{
The \code{recint} function rectifies and integrates electromyography data in a single function. The function takes the absolute value of the EMG trace and then applies a moving average filter to the rectified trace to integrate the signal.  Additionally, it returns a plot of the rectified and integrated EMG signal.
}
\examples{
\dontrun{
# Example usage:
data(EMG_data)

recint_output <- recint(
  input_data = EMG_data,
  window = 25,
  fs = 5000,
  time_units = 'ms',
  preview_plot = TRUE
)
}
}
\author{
Ethan Benevides
}
